requirements.txt
flask
requests
pyswisseph
pytz
waitress